requirements.txt
firecrawl-py
pydantic
jinja2
